refactor: cleanup literal Display trait
Moves display of literal to prql_ast crate.
This code is quite inherent to Literal, so it can reside in prql_ast.
Especially because it simplifies other parts of the codebase.
